Move all ProcessorOptions code to Engine. Convert POR implementations to internal.
