+++ src/Microsoft.Performance.Toolkit.Engine.Tests/ProcessingOptionsResolverTests.cs
@@ -3,7 +3,6 @@
 
 using Microsoft.Performance.SDK.Processing;
 using Microsoft.Performance.SDK.Processing.DataSourceGrouping;
-using Microsoft.Performance.SDK.Processing.Options;
 using Microsoft.Performance.SDK.Runtime;
 using Microsoft.Performance.Testing;
 using Microsoft.Performance.Toolkit.Engine.Tests.TestCookers.Source123;
@@ -90,57 +89,6 @@ namespace Microsoft.Performance.SDK.Tests
             Assert.AreEqual(processorOptions, sut.GetProcessorOptions(dsg123, psr123.Instance), "ProcessorOptions differ from expected");
             Assert.AreEqual(processorOptions, sut.GetProcessorOptions(dsg4, psr4.Instance), "ProcessorOptions differ from expected");
             Assert.AreEqual(processorOptions, sut.GetProcessorOptions(dsg5, psr5.Instance), "ProcessorOptions differ from expected");
-        }
-
-        [TestMethod]
-        [IntegrationTest]
-        public void GlobalProcessingOptionsResolverTestFailure()
-        {
-            var processorOptions = new ProcessorOptions(
-                new[]
-                {
-                    new OptionInstance(
-                        new Option('t', "test2"),
-                        "arg2"),
-                });
-
-            var dsg123 = new DataSourceGroup(
-                new[]
-                {
-                    new FileDataSource("test1" + Source123DataSource.Extension),
-                    new FileDataSource("test2" + Source123DataSource.Extension),
-                    new FileDataSource("test3" + Source123DataSource.Extension)
-                }, new DefaultProcessingMode());
-
-            var dsg4 = new DataSourceGroup(
-                new[]
-                {
-                    new FileDataSource("test4" + Source4DataSource.Extension),
-                }, new DefaultProcessingMode());
-
-            var sut = new GlobalProcessingOptionsResolver(processorOptions);
-
-            Assert.IsNotNull(sut, "Options Resolver is null");
-
-            ProcessingSourceReference.TryCreateReference(typeof(Source123DataSource), out var psr123);
-            ProcessingSourceReference.TryCreateReference(typeof(Source4DataSource), out var psr4);
-
-            // success
-            Assert.AreEqual(processorOptions, sut.GetProcessorOptions(dsg4, psr4.Instance), "ProcessorOptions differ from expected");
-
-            // Failure Expected since all options are not supported
-            try
-            {
-                var actualOptions = sut.GetProcessorOptions(dsg123, psr123.Instance);
-            }
-            catch (NotSupportedException e)
-            {
-                Assert.IsNotNull(e);
-            }
-            catch (Exception e)
-            {
-                Assert.Fail(e.Message);
-            }
         }
 
         [TestMethod]
